chore: renamed views to include View at the end of the file, updated router
