+++ src/views/KillerPerksView.vue
