+++ src/views/KillerView.vue
