+++ src/views/SurvivorPerksView.vue
