+++ src/views/SurvivorView.vue
